\name{pumadyn}
\alias{pumadyn_isample_x}
\alias{pumadyn_isample_y}
\alias{pumadyn_osample_x}
\alias{pumadyn_osample_y}
\title{Pumadyn dataset}
\description{This is a dataset synthetically generated from a realistic simulation of the dynamics of a Unimation Puma 560 robot arm.
} 
\usage{data(pumadyn)}
\format{Response variable vector and design matrix on 4499 in-sample and 3693 out-sample observations for 32 predictor variables.}
\references{
Corke, P. I. (1996). A Robotics Toolbox for MATLAB. \emph{IEEE Robotics and Automation Magazine}, 3 (1): 24-32. 
} 
\keyword{datasets}
