\name{mich_lung}
\alias{mich_lung_xx}
\alias{mich_lung_yy}
\title{Gene expression data for Michigan lung cancer study in Beer et al. (2002)}
\description{Gene expression data of 5217 genes for n = 86 subjects,
with 62 subjects in "good outcomes" (class 1) and 24 subjects in "poor outcomes" (class 2), from the microarray study of Beer et al. (2002).
} 
\usage{data(mich_lung)}
\format{Response variable vector and design matrix on 86 observations for expression of 5217 genes.}
\references{
Beer et al. (1999) Gene-expression profiles predict survival of patients with lung adenocarcinoma. \emph{Nature medicine}, 286(8): 816-824.
} 
\keyword{datasets}
